Warnings

+++ testLab6.c
@@ -1,8 +1,9 @@
 #include "testLab.h"
 #include <stdio.h>
+#include <stdlib.h>
 #include <string.h>
 
-static int ptr_size(void )
+static size_t ptr_size(void )
 {
     static int x64 = -1;
     if (x64 == -1)
@@ -10,10 +11,10 @@ static int ptr_size(void )
         char *cpu = getenv("PROCESSOR_IDENTIFIER");
         x64 = cpu ? strstr(cpu, "x86") == NULL : 1;
     }
-    return sizeof(int)*(x64+1);
+    return sizeof(int)*(size_t)(x64+1);
 }
 
-static int up16(int n)
+static size_t up16(size_t n)
 {
     return (1+(n-1)/16)*16;
 }
@@ -193,7 +194,7 @@ static int checkerBig(void)
 static int feederBig1(void)
 {
     FILE *const in = fopen("in.txt", "w+");
-    int i;
+    unsigned i;
     if (!in) {
         printf("can't create in.txt. No space on disk?\n");
         return -1;
